requirements.txt
fastapi
uvicorn
python-keycloak
python-jose
requests
